Read The Description For Your Own Sanity
So there's a lot to cover in what I've made here. You can download the video of a chosen link.
You can grab the name of a video of a chosen Link
You can show all the graphs and semi-split audio
You can play audio in a pygame terminal
Made a new audio file for testing purposes
Deleted a lot of unnecessary code that I wrote. Don't worry about that.